+++ Changelog.md
@@ -2,9 +2,12 @@
 
 __New Features__
 - Updates to OpenStudio 3.2.1/EnergyPlus 9.5.0.
+- **Breaking change**: Replaces "Unmet Load" outputs with "Unmet Hours".
+- **Breaking change**: Renames "Load: Heating" and "Peak Load: Heating" (and Cooling) outputs to include "Delivered".
 - **Breaking change**: Replaces ClothesDryer `extension/IsVented` and `extension/VentedFlowRate` with `Vented` and `VentedFlowRate`.
 - Expanded capabilities for scheduling:
   - Allows modeling detailed occupancy via a schedule CSV file.
+  - Introduces a measure for automatically generating detailed smooth/stochastic schedule CSV files.
   - Expands simplified weekday/weekend/monthly schedule inputs to additional building features.
   - Allows `HeatingSeason` & `CoolingSeason` to be specified for defining heating and cooling equipment availability.
 - Allows non-zero refrigerant charge defect ratios for ground source heat pumps.
@@ -32,9 +35,11 @@ __New Features__
 - Revises shared mechanical ventilation preconditioning control logic to operate less often.
 - Removes error-check for number of bedrooms based on conditioned floor area, per RESNET guidance.
 - Updates the reporting measure to register all outputs from the annual CSV with the OS runner (for use in, e.g., PAT).
+- Removes timeseries CSV output columns that are all zeroes to reduce file size and processing time.
 - Use constant Cutler curve coefficients in installation quality calculations instead of coil curve coefficients for two-speed, variable-speed systems and ground source heat pump.
-- Fix a bug in setting ground source heat pump rated airflow.
+- Fix a bug where gshp rated airflow is set to actual airflow.
 - Adds an `--ep-input-format` argument to run_simulation.rb to choose epJSON as the EnergyPlus input file format instead of IDF.
+- Eliminates EnergyPlus warnings related to unused objects or invalid output meters/variables.
 
 __Bugfixes__
 - Improves ground reflectance when there is shading of windows/skylights.
@@ -42,6 +47,7 @@ __Bugfixes__
 - Fixes mechanical ventilation compartmentalization area calculation for SFA/MF homes with surfaces with InteriorAdjacentTo==ExteriorAdjacentTo.
 - Negative `DistanceToTopOfInsulation` values are now disallowed.
 - Fixes workflow errors if a `VentilationFan` has zero airflow rate or zero hours of operation.
+- Fixes duct design load calculations for HPXML files with multiple ducted HVAC systems.
 
 ## OpenStudio-HPXML v1.2.0